Makes server mem info command configurable.
It might be useful to have the server monitor memory on another
system. It could monitor an Android device with
`adb shell cat /proc/meminfo`, or a remote system with
`ssh cloudserver cat /proc/meminfo`.

Not sure if it's safe to allow an environment variable to specify
a command for the server to execute once per second. Going with it
for now, but it might make sense to make this a shell script, or
make the server all the command with `spawn`, which would require
that the command is a file, not a shell command like `cat`.